split up nml_upload_zone_container.tsx module
